azure-ml-pipeline/requirements.txt
azureml-sdk
azureml-pipeline-core 
azureml-pipeline-steps 
pandas 
requests